MLPAB-2607 revert setup
